16. In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -1,31 +1,32 @@
 #include <stdio.h>
 /**
- * main - prints the first 98 Fibonacci numbers, starting 
- * with 1 and 2, followed by new line
+ * main - finds and prints the first 98 Fibonacci numbers,
+ * starting with 1 and 2, followed by a new line.
  *
  * Return: 0
  */
 
 int main(void)
 {
-	unsigned long long fib1 = 1;
-	unsigned long long fib2 = 2;
-	unsigned long long fib3;
+	unsigned int fib1 = 1;
+	unsigned int fib2 = 2;
+	unsigned int fib3;
 	int count = 2;
 
-	printf("%llu, %llu, ", fib1, fib2);
-	while (count < 98) {
+	printf("%u, %u, ", fib1, fib2);
+	while (count < 98)
+	{
 		fib3 = fib1 + fib2;
-		printf("%llu", fib3);
-        /*Add comma and space except for the last number*/
+		printf("%u", fib3);
+	/* Add comma and space except for the last number*/
 	if (count < 97)
 	{
-		printf(", ");
+	printf(", ");
 	}
 	fib1 = fib2;
 	fib2 = fib3;
 	count++;
-}
+	}
 	printf("\n");
 return 0;
 }
